feat: validate add key popup
